# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Trading Buddy API",
    "description": "A股量化交易数据基础设施",
    "version": "1.1.0"
  },
  "paths": {
...
    "/api/backtest/ma-cross/signal": {
      "get": {
        "tags": [
          "回测"
        ],
        "summary": "Ma Cross Signal",
        "description": "最近一根有效 K 上的双均线多空状态（无权益曲线、无费率）。",
        "operationId": "ma_cross_signal_api_backtest_ma_cross_signal_get",
        "parameters": [
          {
            "name": "code",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "description": "标的代码",
              "title": "Code"
            },
            "description": "标的代码"
          },
          {
            "name": "fast",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 120,
              "minimum": 1,
              "default": 5,
              "title": "Fast"
            }
          },
          {
            "name": "slow",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 500,
              "minimum": 2,
              "default": 20,
              "title": "Slow"
            }
          },
          {
            "name": "start_date",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "date"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Start Date"
            }
          },
          {
            "name": "end_date",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "date"
                },
                {
                  "type": "null"
                }
              ],
              "title": "End Date"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 5000,
              "minimum": 30,
              "description": "与单标的回测一致；须 ≥ slow",
              "default": 500,
              "title": "Limit"
            },
            "description": "与单标的回测一致；须 ≥ slow"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MaCrossSignalResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/watchlist/items": {
      "get": {
        "tags": [
          "自选"
        ],
        "summary": "Watchlist Items",
        "operationId": "watchlist_items_api_watchlist_items_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WatchlistItemsResponse"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "自选"
        ],
        "summary": "Watchlist Add Item",
        "operationId": "watchlist_add_item_api_watchlist_items_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/WatchlistAddRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "type": "object",
                  "title": "Response Watchlist Add Item Api Watchlist Items Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/watchlist/items/{code}": {
      "delete": {
        "tags": [
          "自选"
        ],
        "summary": "Watchlist Remove Item",
        "operationId": "watchlist_remove_item_api_watchlist_items__code__delete",
        "parameters": [
          {
            "name": "code",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Code"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Watchlist Remove Item Api Watchlist Items  Code  Delete"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/strategies/signal": {
      "post": {
        "tags": [
          "策略"
        ],
        "summary": "Post Strategy Signal",
        "description": "用 JSON 模板请求最新策略信号（当前仅 `ma_cross`）。\n与 `GET /api/backtest/ma-cross/signal` 语义一致，便于自动化与后续多策略扩展。",
        "operationId": "post_strategy_signal_api_strategies_signal_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/StrategySignalRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StrategySignalResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "MaCrossSignalResponse": {
        "properties": {
          "code": {
            "type": "string",
            "title": "Code"
          },
          "fast_period": {
            "type": "integer",
            "title": "Fast Period"
          },
          "slow_period": {
            "type": "integer",
            "title": "Slow Period"
          },
          "bars_used": {
            "type": "integer",
            "title": "Bars Used"
          },
          "as_of_date": {
            "type": "string",
            "title": "As Of Date"
          },
          "position": {
            "type": "string",
            "title": "Position",
            "description": "long 或 flat（收盘后 MA 比较）"
          },
          "close": {
            "type": "number",
            "title": "Close"
          },
          "ma_fast": {
            "type": "number",
            "title": "Ma Fast"
          },
          "ma_slow": {
            "type": "number",
            "title": "Ma Slow"
          },
          "note": {
            "type": "string",
            "title": "Note"
          }
        },
        "type": "object",
        "required": [
          "code",
          "fast_period",
          "slow_period",
          "bars_used",
          "as_of_date",
          "position",
          "close",
          "ma_fast",
          "ma_slow",
          "note"
        ],
        "title": "MaCrossSignalResponse"
      },
...
      "StrategySignalRequest": {
        "properties": {
          "code": {
            "type": "string",
            "maxLength": 64,
            "minLength": 1,
            "title": "Code",
            "description": "标的代码"
          },
          "kind": {
            "type": "string",
            "maxLength": 64,
            "minLength": 1,
            "title": "Kind",
            "default": "ma_cross"
          },
          "params": {
            "additionalProperties": true,
            "type": "object",
            "title": "Params"
          }
        },
        "type": "object",
        "required": [
          "code"
        ],
        "title": "StrategySignalRequest",
        "description": "POST /api/strategies/signal 请求体；params 按 kind 解析。"
      },
      "StrategySignalResponse": {
        "properties": {
          "kind": {
            "type": "string",
            "title": "Kind"
          },
          "signal": {
            "$ref": "#/components/schemas/MaCrossSignalResponse"
          }
        },
        "type": "object",
        "required": [
          "kind",
          "signal"
        ],
        "title": "StrategySignalResponse"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
      "WatchlistAddRequest": {
        "properties": {
          "code": {
            "type": "string",
            "maxLength": 24,
            "minLength": 4,
            "title": "Code",
            "description": "标的代码，如 sh.600000"
          }
        },
        "type": "object",
        "required": [
          "code"
        ],
        "title": "WatchlistAddRequest"
      },
      "WatchlistItemOut": {
        "properties": {
          "code": {
            "type": "string",
            "title": "Code"
          },
          "name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Name"
          },
          "created_at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Created At"
          }
        },
        "type": "object",
        "required": [
          "code"
        ],
        "title": "WatchlistItemOut"
      },
      "WatchlistItemsResponse": {
        "properties": {
          "watchlist_id": {
            "type": "integer",
            "title": "Watchlist Id"
          },
          "items": {
            "items": {
              "$ref": "#/components/schemas/WatchlistItemOut"
            },
            "type": "array",
            "title": "Items"
          }
        },
        "type": "object",
        "required": [
          "watchlist_id",
          "items"
        ],
        "title": "WatchlistItemsResponse"
      }
    }
  }
}